server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psutil==5.9.6
python-dotenv==1.0.0
nvidia-ml-py==11.525.112 